More efficient getRecordSize()
which allows getting record size to be inlined in CommonAbstractStore.
